requirements.txt
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1
Flask-Cors==3.0.10
openai==0.27.4
python-dotenv==1.0.0
easygoogletranslate==0.0.4
firebase-admin==6.1.0